ADD Knife Backstab Convar and code
Requested by Add, If a player is looking the same direction (About) as another player, they will "Backstab" with the knife for an instant kill. Otherwise, they only stab for a measly 50 damage. Does not work unless the convar "ttt_backstabs_on" is true. Else knives instantkill all the time.